% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doRandomSearch.R
\name{doRandomSearch}
\alias{doRandomSearch}
\title{Run a random search.}
\usage{
doRandomSearch(fn, par.set, control, ...)
}
\arguments{
\item{fn}{[\code{\link[smoof]{smoof_single_objective_function}}]\cr
Objective function.}

\item{par.set}{[\code{\link[ParamHelpers]{ParamSet}}]\cr
Param set.}

\item{control}{[\code{\link{FocusSearchControl}}]\cr
Control object.}

\item{\ldots}{[any]\cr
Passed to \code{fn}.}
}
\value{
[\code{list}]. 
  \item{x [list]}{s}
  \item{y [numeric]}{s}
}
\description{
fooo
}
