% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FocusSearchControl.R
\name{makeFocusSearchControl}
\alias{makeFocusSearchControl}
\alias{FocusSearchControl}
\title{Create control object for focus search.}
\usage{
makeFocusSearchControl(points = 100L, maxit = 1L, restarts = 1L,
  exploit = 1)
}
\arguments{
\item{points}{[\code{integer(1)}]\cr
Number of points for each random search.
Default is 100.}

\item{maxit}{[\code{integer(1)}]\cr
Number of focus interation where feasible region is shrinked.
Default is 100.}

\item{restarts}{[\code{integer(1)}]\cr
Number of independent restarts.
Default is 1.
@param exploit (`numeric`) \¢r
Fraction of points to exploit in the local param space vs. in the global param space.
Defaults to 1 (full exploitation).}
}
\value{
[\code{\link{FocusSearchControl}}]
}
\description{
Creates a control object to be used with (`focussearch`). 
See __Arguments__ for further details.
}
