% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkParSet.R
\name{shrinkParSet}
\alias{shrinkParSet}
\title{Shrink param set towards a point.}
\usage{
shrinkParSet(par.set, x.df)
}
\arguments{
\item{par.set}{[\code{\link[ParamHelpers]{ParamSet}}]\cr
Param set.}
}
\value{
[\code{\link{ParamSet}}]
}
\description{
fooo
}
